Added re-branch functionality if the local bazaar repo is broken/corrputed
When you have a lot's of slaves and big repos (>2G), the branch took
several minutes. If we misconfigured the build timeout or we're just
inpatient, it can be easily aborted. In this case, our bazaar
repository will be corrupted. So we have to log in to the related
jenkins node and delete the workspace manually.

With this patch, it will be automatic.