prevent nullptrexceptions from accessing baseline when polling
baseline is a previous revision generated by one of our functions, if that
came out empty, or the job plugin has a bug, or the data got corrupted it
may be null. to prevent exceptions in this case, simply check for nullness
before accessing the object.

when baseline is null we always indicate a significant change.