fix lightweight checkout revision detection
for lightweight branches bzr revision-info would query the remote to get
the latest revision in the remote branch rather than the local lightweight
version. --tree prevents this from happening and always reports the
revision of the tree at `root` regardless of the branch type at root.

since revision-info is used both for local and remote queries there's a
new helper getLocalRevisionState() which will add --tree while the default
behavior is remote querying which continues to not use --tree.

this fixes lightweight branches to actually poll properly